sync - circular
